src/main/java/domain/LottoGame.java
package domain;

import java.util.ArrayList;
import java.util.List;

public class LottoGame {
    private final List<Lotto> lottos;

    public LottoGame(List<Lotto> lottos) {
        this.lottos = lottos;
    }

    public static LottoGame create(int lottoCount) {
        List<Lotto> lottos = new ArrayList<>();

        for (int i = 0; i < lottoCount; i++) {
            lottos.add(new Lotto());
        }

        return new LottoGame(lottos);
    }

    public List<Lotto> getLottos() {
        return lottos;
    }
}


